fs: PR feedback to unify Files settings.
