--- shared/settings/files/index.native.js
@@ -1,5 +1,0 @@
-// @flow
-import * as React from 'react'
-import {Box} from '../../common-adapters'
-
-export default () => <Box />
